+++ assets/src/components/layout/AppLayout.tsx
@@ -2,11 +2,11 @@ import { Fragment, RenderableProps } from 'preact'
 import { route } from 'preact-router'
 import { useEffect, useState } from 'preact/hooks'
 
-import AppSidebar from '@/components/layout/AppSidebar'
-import UserMenu from '@/components/layout/UserMenu'
 import TaskContextMenu from '@/components/menus/TaskContextMenu'
 import DeleteTaskModal from '@/components/modals/DeleteTaskModal'
 import QuickFindModal from '@/components/modals/QuickFindModal'
+import AppSidebar from '@/components/nav/AppSidebar'
+import UserMenu from '@/components/nav/UserMenu'
 import ProjectPills from '@/components/projects/ProjectPills'
 import { paths } from '@/config'
 import useSwipe from '@/hooks/useSwipe'
